added a delete option and set max steps

+++ app.py
@@ -5,7 +5,7 @@ import uuid, OpenSSL
 import requests
 from flask import Flask, request
 from flask_sqlalchemy import SQLAlchemy
-from wit import Wit 
+from wit import Wit
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ['DATABASE_URL']
@@ -21,8 +21,8 @@ name = None
 date = None
 email = None
 sender_id = None
-appointee = None
- 
+# appointee = None
+
 @app.route('/', methods=['GET'])
 def verify():
     # when the endpoint is registered as a webhook, it must echo back
@@ -31,46 +31,46 @@ def verify():
         if not request.args.get("hub.verify_token") == os.environ["VERIFY_TOKEN"]:
             return "Verification token mismatch", 403
         return request.args["hub.challenge"], 200
- 
+
     return "Hello world", 200
- 
- 
+
+
 @app.route('/', methods=['POST'])
 def webhook():
- 
+
      # endpoint for processing incoming messaging events
- 
+
     data = request.get_json()
     log(data)  # you may not want to log every incoming message in production, but it's good for testing
- 
+
     if data["object"] == "page":
- 
+
         for entry in data["entry"]:
             for messaging_event in entry["messaging"]:
- 
+
                 if messaging_event.get("message"):  # someone sent us a message
                     global sender_id
                     sender_id = messaging_event["sender"]["id"]        # the facebook ID of the person sending you the message
                     recipient_id = messaging_event["recipient"]["id"]  # the recipient's ID, which should be your page's facebook ID
                     message_text = messaging_event["message"]["text"]  # the message's text
- 
+
                     curr_context = {}
-                    client.run_actions(sender_id, message_text, curr_context)
- 
+                    client.run_actions(sender_id, message_text, curr_context, 25)
+
                 if messaging_event.get("delivery"):  # delivery confirmation
                     pass
- 
+
                 if messaging_event.get("optin"):  # optin confirmation
                     pass
- 
+
                 if messaging_event.get("postback"):  # user clicked/tapped "postback" button in earlier message
                     pass
- 
+
     return "ok", 200
- 
- 
+
+
 def send_message(recipient_id, message_text):
- 
+
     log("sending message to {recipient}: {text}".format(recipient=recipient_id, text=message_text))
 
     params = {
@@ -85,18 +85,18 @@ def send_message(recipient_id, message_text):
         },
         "message": {
             "text": message_text
-           # "quick_replies":[
-           #                     {
-           #                      "content_type":"text",
-           #                       "title":"Yes",
-           #                       "payload":"DEVELOPER_DEFINED_PAYLOAD_FOR_PICKING_GREEN"
-           #                       },
-           #                      {
-           #                        "content_type":"text",
-           #                        "title":"No",
-           #                        "payload":"DEVELOPER_DEFINED_PAYLOAD_FOR_PICKING_RED"
-           #                         }
-           #                 ]
+        #    "quick_replies":[
+        #                        {
+        #                         "content_type":"text",
+        #                          "title":"Yes",
+        #                          "payload":"DEVELOPER_DEFINED_PAYLOAD_FOR_PICKING_GREEN"
+        #                          },
+        #                         {
+        #                           "content_type":"text",
+        #                           "title":"No",
+        #                           "payload":"DEVELOPER_DEFINED_PAYLOAD_FOR_PICKING_RED"
+        #                            }
+        #                    ]
 
         }
     })
@@ -104,12 +104,12 @@ def send_message(recipient_id, message_text):
     if r.status_code != 200:
         log(r.status_code)
         log(r.text)
- 
- 
+
+
 def log(message):  # simple wrapper for logging to stdout on heroku
     print str(message)
     sys.stdout.flush()
- 
+
 def first_entity_value(entities, entity):
     """
     Returns first entity value
@@ -120,7 +120,7 @@ def first_entity_value(entities, entity):
     if not val:
         return None
     return val['value'] if isinstance(val, dict) else val
- 
+
 def add_appointment(request):
     context = request['context']
     entities = request['entities']
@@ -140,20 +140,20 @@ def add_appointment(request):
         if context.get('date') is not None:
             del context['date']
     return context
- 
+
 def parse_datetime(datetime):
     date_array = datetime[0:datetime.index('T')].split('-')
-    date = str(date_array[1]) + '/' + str(date_array[2]) + '/' + str(date_array[0]) 
+    date = str(date_array[1]) + '/' + str(date_array[2]) + '/' + str(date_array[0])
     return date
- 
+
 def get_events(request):
     context = request['context']
     page_access_token = os.environ['PAGE_ACCESS_TOKEN']
- 
+
     result = requests.get('https://graph.facebook.com/v2.8/dummyanvilpage/events?access_token=' + page_access_token).json()
- 
+
     data = result['data']
- 
+
     if data:
         message = ''
         for event in data:
@@ -162,14 +162,14 @@ def get_events(request):
             event_id = event['id']
             message = message + 'Name: ' + event_name +  '\nDescription:\n' + event_description + '\nLink: ' + 'https://www.facebook.com/events/' + event_id
             message = message + "\n\n"
-         
+
         message = message[:len(message) - 2]
-        context['event'] = message       
+        context['event'] = message
     else:
         context['event'] = 'Sorry there are no upcoming events!'
-     
+
     return context
- 
+
 def get_user_info():
     page_access_token = os.environ['PAGE_ACCESS_TOKEN']
     result = requests.get('https://graph.facebook.com/v2.8/' + sender_id + '?fields=first_name,last_name&access_token=' + page_access_token).json()
@@ -177,28 +177,41 @@ def get_user_info():
     last_name = result['last_name']
     name = first_name + ' ' + last_name
     return name
-  
+
 def get_email(request):
     entities = request['entities']
     global email
     email = first_entity_value(entities, 'email')
     update_db(request)
     return request['context']
-  
+
 def send(request, response):
     send_message(request['session_id'], response['text'])
- 
+
 def update_db(request):
    # print "update_db() vals:\nname: " + str(user_name) + "\nemail: " + str(email) + "\ndate: " + str(appointment_date)
-    global appointee
-    appointee = AnvilAppointment(name, email, date)
-    db.session.add(appointee)
-    db.session.commit()
+    # global appointee
+    appointee = AnvilAppointment.query.filter_by(name=name).first()
+    if (appointee):
+        appointee.name = name
+        appointee.email = email
+        appointee.appointment_date = date
+        db.session.commit()
+    else:
+        appointee = AnvilAppointment(name, email, date)
+        db.session.add(appointee)
+        db.session.commit()
 
-def delete_apt(request): 
-    db.session.delete(appointee)
-    db.session.commit()
-    return request['context'] 
+def delete_apt(request):
+    context = request['context'];
+    appointee = AnvilAppointment.query.filter_by(name=name).first()
+    if (appointee):
+        db.session.delete(appointee)
+        db.session.commit()
+        context['success'] = "Your appointment was deleted successfully!"
+    else:
+        context['success'] = "We were not able to delete your appointment."
+    return request['context']
 
 actions = {
  'send' : send,
@@ -207,8 +220,8 @@ actions = {
   'get_email' : get_email,
   'delete_apt' : delete_apt,
  }
- 
+
 client = Wit(access_token=access_token, actions=actions)
- 
+
 if __name__ == '__main__':
     app.run(debug=True)
